退会機能の作成
