+++ src/style/atoms/toppage/Sidebar.scss
@@ -16,6 +16,12 @@ aside {
                 margin-bottom: 0;
             }
         }
+
+        p.withdrawal {
+            font-size: 1.4rem;
+            margin-bottom: 0;
+            cursor: pointer;
+        }
     }
 
     button {
